+++ app/src/main/java/com/example/hideandseek/ui/view/ResultScreen.kt
